# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Device {
  id: string;
  code: string;
  displayName: string;
  online: boolean;
  createdAt: string;
}

export interface DeviceInput {
  /** @maxLength 60 */
  displayName?: string;
}

export interface PresenceInput {
  online: boolean;
}

export interface Message {
  id: string;
  chatId: string;
  senderId: string;
  /** @nullable */
  senderCode?: string | null;
  text: string;
  createdAt: string;
}

export type ChatType = typeof ChatType[keyof typeof ChatType];


export const ChatType = {
  direct: 'direct',
  group: 'group',
} as const;

export type ChatStatus = typeof ChatStatus[keyof typeof ChatStatus];


export const ChatStatus = {
  pending_incoming: 'pending_incoming',
  pending_outgoing: 'pending_outgoing',
  active: 'active',
  rejected: 'rejected',
} as const;

export interface ChatMember {
  deviceId: string;
  code: string;
  displayName: string;
  online: boolean;
}

export interface Chat {
  id: string;
  type: ChatType;
  name: string;
  status: ChatStatus;
  members: ChatMember[];
  messages: Message[];
  updatedAt: string;
}

export type ChatRequestStatus = typeof ChatRequestStatus[keyof typeof ChatRequestStatus];


export const ChatRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface ChatRequest {
  id: string;
  chatId: string;
  fromDeviceId: string;
  toDeviceId: string;
  fromCode: string;
  toCode: string;
  status: ChatRequestStatus;
  createdAt: string;
}

export interface DeviceSync {
  device: Device;
  chats: Chat[];
  requests: ChatRequest[];
}

export interface ChatRequestInput {
  /**
     * @minLength 4
     * @maxLength 12
     */
  targetCode: string;
}

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 25
     * @items.minLength 4
     * @items.maxLength 12
     */
  memberCodes: string[];
}

export type RequestResponseInputStatus = typeof RequestResponseInputStatus[keyof typeof RequestResponseInputStatus];


export const RequestResponseInputStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface RequestResponseInput {
  deviceId: string;
  status: RequestResponseInputStatus;
}

export interface MessageInput {
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
}