# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMember } from './chatMember';
import type { ChatStatus } from './chatStatus';
import type { ChatType } from './chatType';
import type { Message } from './message';

export interface Chat {
  id: string;
  type: ChatType;
  name: string;
  status: ChatStatus;
  members: ChatMember[];
  messages: Message[];
  updatedAt: string;
}