# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestStatus } from './chatRequestStatus';

export interface ChatRequest {
  id: string;
  chatId: string;
  fromDeviceId: string;
  toDeviceId: string;
  fromCode: string;
  toCode: string;
  status: ChatRequestStatus;
  createdAt: string;
}