# lib/api-zod/src/generated/types/chatRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequestInput {
  /**
     * @minLength 4
     * @maxLength 12
     */
  targetCode: string;
}