# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInput {
  /** @maxLength 60 */
  displayName?: string;
}