# lib/api-zod/src/generated/types/deviceSync.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Chat } from './chat';
import type { ChatRequest } from './chatRequest';
import type { Device } from './device';

export interface DeviceSync {
  device: Device;
  chats: Chat[];
  requests: ChatRequest[];
}