# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 25
     * @items.minLength 4
     * @items.maxLength 12
     */
  memberCodes: string[];
}