# lib/api-zod/src/generated/types/requestResponseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestResponseInputStatus } from './requestResponseInputStatus';

export interface RequestResponseInput {
  deviceId: string;
  status: RequestResponseInputStatus;
}